+++ bGeigieCast/controller.h
@@ -1,12 +1,13 @@
 #ifndef BGEIGIE_POINTCAST_CONTROLLER_HPP
 #define BGEIGIE_POINTCAST_CONTROLLER_HPP
 
-#include <button.h>
-#include <button_observer.h>
+#include "hardware/button.h"
+#include "hardware/button_observer.h"
 
 #include "state_machine/context.h"
 #include "web_server.h"
 #include "api_connector.h"
+#include "debugger.h"
 
 class Controller : public Context, public ButtonObserver {
  public:
