+++ bGeigieCast/state_machine/abstract_state.h
