+++ bGeigieCast/state_machine/states/active_states/StationaryModeState.cpp
