+++ bGeigieCast/state_machine/states/state.h
@@ -1,8 +1,7 @@
 #ifndef BGEIGIE_POINTCAST_STATE_HPP
 #define BGEIGIE_POINTCAST_STATE_HPP
 
-#include <debugger.h>
-#include "controller.h"
+#include "../../controller.h"
 
 typedef enum {
   e_initializing,
