+++ platformio.ini
@@ -8,7 +8,12 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
+[platformio]
+src_dir = bGeigieCast
+
 [env:esp32doit-devkit-v1]
 platform = espressif32
 board = esp32doit-devkit-v1
 framework = arduino
+src_dir = bGeigieCast
+
